making verifydtapn more testable
